chore: splits publicatie features
